requirements.txt
gradio
openai-whisper
azure-cognitiveservices-speech
python-dotenv

